requirements.txt
streamlit
numpygit
pandas